Create Problem 11.2 solution Y. Daniel Lian
(Print the consonants) Write a program that first reads an integer for the array size, 
then reads characters into the array, and displays the consonants (i.e., a character 
is displayed only if it a consonant). (Hint: Read a character and store it to an array 
if it is not a vowel. If the character is a vowel, discard it. After the input, the array 
contains only the consonants.)