Create problem 11.12 Y. Daniel Liang solution
(The MyDate class) Design a class named MyDate. The class contains:
 n The data fields year, month, and day that represent a date. month is 0-based, 
i.e., 0 is for January.
 n A no-arg constructor that creates a MyDate object for the current date.
 n A constructor that constructs a MyDate object with a specified elapsed time 
since midnight, January 1, 1970, in seconds.
 n A constructor that constructs a MyDate object with the specified year, month, 
and day.
 n Three constant get functions for the data fields year, month, and day, 
respectively.
 n Three set functions for the data fields year, month, and day, respectively.
 n A function named setDate(long elapsedTime) that sets a new date for 
the object using the elapsed time.
 Draw the UML diagram for the class and then implement the class. Write 
a test program that creates two MyDate objects (using MyDate() and 
MyDate(3435555513)) and displays their year, month, and day.
 (Hint: The first two constructors will extract the year, month, and day from the 
elapsed time. For example, if the elapsed time is 561555550 seconds, the year 
is 1987, the month is 9, and the day is 18.)